refactor : notion Util name
